+++ src/components/Nav/Reply.js
@@ -1,22 +1,7 @@
 import './Reply.scss';
-import { useState, useEffect } from 'react';
 
-function Reply() {
-  const [reply, setReply] = useState('');
-  const [replies, setReplies] = useState([]);
-  const onChange = event => setReply(event.target.value);
-  const onSubmit = event => {
-    event.preventDefault();
-    if (reply === '') {
-      return;
-    }
-    setReplies(currentArray => [...currentArray, reply]);
-    setReply('');
-  };
-  const deleteBtn = event => {
-    const li = event.target.parentElement;
-    li.remove();
-  };
+function Reply(props) {
+  const {replies, deleteBtn } = props;
   return (
     <>
       <ul className="reply">
@@ -30,17 +15,7 @@ function Reply() {
           </li>
         ))}
       </ul>
-      <div className="time">42분 전</div>
-      <form className="reply__input" onSubmit={onSubmit}>
-        <input
-          onChange={onChange}
-          value={reply}
-          className="reply__input__text"
-          type="text"
-          placeholder="댓글 달기..."
-        />
-        <button className="reply__input__btn">게시</button>
-      </form>
+     
     </>
   );
 }
