+++ src/pages/YoonJi/Main/MainYoonJi.js
@@ -5,6 +5,21 @@ import Reply from '../../../components/Nav/Reply';
 // import React from "react";
 
 function MainYoonji() {
+  const [reply, setReply] = useState('');
+  const [replies, setReplies] = useState([]);
+  const onChange = event => setReply(event.target.value);
+  const onSubmit = event => {
+    event.preventDefault();
+    if (reply === '') {
+      return;
+    }
+    setReplies(currentArray => [...currentArray, reply]);
+    setReply('');
+  };
+  const deleteBtn = event => {
+    const li = event.target.parentElement;
+    li.remove();
+  };
   return (
     <>
       <nav>
@@ -64,7 +79,18 @@ function MainYoonji() {
                 <span>syj3367</span> 복순이가 탈출해서 이불에 들어옴
               </div>
 
-              <Reply />
+              <Reply replies={replies} deleteBtn ={deleteBtn} />
+              <div className="time">42분 전</div>
+      <form className="reply__input" onSubmit={onSubmit}>
+        <input
+          onChange={onChange}
+          value={reply}
+          className="reply__input__text"
+          type="text"
+          placeholder="댓글 달기..."
+        />
+        <button className="reply__input__btn">게시</button>
+      </form>
             </div>
           </article>
         </section>
